Move JobProxyDispatcherHelper to Spec::Support::JobProxyDispatcherHelper
