+++ spec/models/job_proxy_dispatcher_get_eligible_proxies_for_job_spec.rb
@@ -1,6 +1,6 @@
 describe "JobProxyDispatcherGetEligibleProxiesForJob" do
-  require File.expand_path(File.join(File.dirname(__FILE__), 'job_proxy_dispatcher/job_proxy_dispatcher_helper'))
-  include JobProxyDispatcherHelper
+  include Spec::Support::JobProxyDispatcherHelper
+
   context "with two servers on same zone, vix disk enabled for all, " do
     before(:each) do
       @server1 = EvmSpecHelper.local_miq_server
