+++ spec/models/job_proxy_dispatcher_vm_proxies4job_spec.rb
@@ -1,6 +1,5 @@
 describe "JobProxyDispatcherVmProxies4Job" do
-  require File.expand_path(File.join(File.dirname(__FILE__), 'job_proxy_dispatcher/job_proxy_dispatcher_helper'))
-  include JobProxyDispatcherHelper
+  include Spec::Support::JobProxyDispatcherHelper
 
   context "with two servers on same zone, vix disk enabled for all, " do
     before(:each) do
