+++ spec/models/job_proxy_dispatcher_vm_storage2hosts_spec.rb
@@ -1,6 +1,6 @@
 describe "JobProxyDispatcherVmStorage2Hosts" do
-  require File.expand_path(File.join(File.dirname(__FILE__), 'job_proxy_dispatcher/job_proxy_dispatcher_helper'))
-  include JobProxyDispatcherHelper
+  include Spec::Support::JobProxyDispatcherHelper
+
   context "with two servers on same zone, vix disk enabled for all, " do
     before(:each) do
       @server1 = EvmSpecHelper.local_miq_server(:is_master => true)
